% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhcweek5.R
\docType{package}
\name{zhcweek5-package}
\alias{zhcweek5}
\alias{zhcweek5-package}
\title{zhcweek5: Extract the values for a specific test run by FastQC on a single fastq file}
\description{
You can specify the test result you want to extract such as
  "Per base sequence quality", "Per sequence quality score" etc.
}
\keyword{internal}
