\docType{package}
\name{zhcweek5}
\alias{zhcweek5}
\alias{zhcweek5-package}
\title{zhcweek5: a practice of ANGDS Week5 assignment}
\description{
  There is only one function in this package:
  reading_in
}
\author{
  \strong{Maintainer}: Zhuoyang Chen \email{zhc4002@med.cornell.edu}

  Authors:
    \itemize{
      \item Zhuoyang Chen
    }
}
\keyword{internal}
